+++ test_chords_used.js
@@ -1,0 +1,1 @@
+// Just checking if ChordsFeature is actually used anywhere other than tests.